04.05.22.21.46

+++ src/scripts/Card.js
@@ -26,14 +26,19 @@ export const initialCards = [
 ];
 
 export class Card {
-  constructor({ item, handleCardClick, handleDeleteIconClick}, cardSelector, api) {
+  constructor({ item, handleCardClick, handleDeleteIconClick }, cardSelector) {
     this._name = item.name;
     this._link = item.link;
     this._counter = item.likes.length;
     this._cardSelector = cardSelector;
     this._handleCardClick = handleCardClick;
     this._handleDeleteIconClick = handleDeleteIconClick;
-    this._api = api;
+    this._ownerId = item.owner._id;
+    this._userId = "ca2be27f0f73ca0ef262d045";
+    this._cardElement = this._getTemplate();
+    this._cardImage = this._cardElement.querySelector(".element__image");
+    this._elementCounter = this._cardElement.querySelector(".element__counter");
+    this._elementDelete = this._cardElement.querySelector(".element__delete");
   }
   _getTemplate() {
     const itemTemplateContent = document
@@ -42,18 +47,22 @@ export class Card {
       .cloneNode(true);
     return itemTemplateContent;
   }
-  
+
   createCard() {
-    this._cardElement = this._getTemplate();
-    this._cardImage = this._cardElement.querySelector(".element__image");
-    this._elementCounter = this._cardElement.querySelector(".element__counter");
     this._likeCounter();
+    this._checkId();
     this._setEventListeners();
     this._cardImage.src = this._link;
     this._cardElement.querySelector(".element__title").textContent = this._name;
     this._cardImage.alt = this._name;
-
     return this._cardElement;
+    
+  }
+
+  _checkId() {
+    if (this._ownerId !== this._userId) {
+      this._elementDelete.remove();
+    }
   }
 
   _likeCounter() {
@@ -71,14 +80,16 @@ export class Card {
   // }
 
   _setEventListeners() {
-    this._cardElement
-      .querySelector(".element__delete")
-      .addEventListener("click", 
-      // this._handleDelete
-     () => {
-      this._handleDeleteIconClick()
-     } 
+    if (this._cardElement.querySelector(".element__delete")) {
+      //проверяем есть элемент в DOM
+      this._elementDelete.addEventListener(
+        "click",
+        // this._handleDelete
+        () => {
+          this._handleDeleteIconClick();
+        }
       );
+    }
 
     this._cardElement
       .querySelector(".element__like")
@@ -91,7 +102,6 @@ export class Card {
         // name: this._name,
         // link: this._link,
       });
-     
     });
   }
 }
